requirements.txt
aiohttp==3.8.3
aiodns==3.0.0
python-json-logger==2.0.4
pdfkit-async==0.7.2